+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Unreleased Features
 Please add a note of your changes below this heading if you make a Pull Request.
 
+### Added
+* Communication watchdog feature.
+* `encoder.set_linear_count(count)` function.
+
 # Releases
 ## [0.4.8] - 2019-02-25
 ### Added
@@ -14,6 +18,9 @@ Please add a note of your changes below this heading if you make a Pull Request.
 * Analog input. Used the same way as the PWM input mappings.
 * Voltage limit soft clamping instead of ERROR_MODULATION_MAGNITUDE in gimbal motor closed loop.
 * Thermal current limit with linear derating.
+
+### Changed
+* Unified lockin drive modes. Current for index searching and encoder offset calibration now moved to axis.lockin.current.
 
 ### Fixed
 * Added required 1.5 cycle phase shift between ADC and PWM, lack thereof caused unstable current controller at high eRPM.